# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Scriptoryum API",
    "description": "API para gerenciamento e análise de documentos do Scriptoryum",
    "contact": {
      "name": "Time Scriptoryum",
      "email": "contato@scriptoryum.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Account/profile": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Obter perfil do usuário",
        "description": "Retorna as informações completas do perfil do usuário autenticado",
        "responses": {
          "200": {
            "description": "Perfil do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "Atualizar perfil",
        "description": "Atualiza as informações do perfil do usuário autenticado",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "409": {
            "description": "Conflito - Email ou nome de usuário já existe"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/api/Account/change-password": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Alterar senha",
        "description": "Altera a senha do usuário autenticado",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou senha atual incorreta"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/api/Account/delete-account": {
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Excluir conta",
        "description": "Exclui permanentemente a conta do usuário autenticado",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conta excluída com sucesso"
          },
          "400": {
            "description": "Senha incorreta"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/api/Account/documents": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Listar documentos do usuário",
        "description": "Retorna todos os documentos associados ao usuário autenticado",
        "responses": {
          "200": {
            "description": "Lista de documentos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordDto": {
        "required": [
          "confirmNewPassword",
          "currentPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "maxLength": 100,
            "minLength": 6,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$",
            "type": "string"
          },
          "confirmNewPassword": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProfileDto": {
        "required": [
          "email",
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          }
        },
        "additionalProperties": false
      },
...
      "UserInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header usando o esquema Bearer. \r\n\r\nDigite 'Bearer' [espaço] e então seu token na entrada de texto abaixo.\r\n\r\nExemplo: \"Bearer 12345abcdef\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}